refactor: fee stats component
